# packages/contracts/openapi/api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/businesses/{id}/reviews": {
      "get": {
        "operationId": "ReviewsPublicController_list",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "newest",
              "type": "string",
              "enum": [
                "newest",
                "highest",
                "lowest"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicReviewDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Approved reviews for a published business, newest first",
        "tags": [
          "public-reviews"
        ]
      },
      "post": {
        "operationId": "ReviewsPublicController_submit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitReviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionReceiptDto"
                }
              }
            }
          }
        },
        "summary": "Submit a review for moderation (Idempotency-Key required)",
        "tags": [
          "public-reviews"
        ]
      }
    },
    "/api/v1/reports": {
      "post": {
        "operationId": "ReviewsPublicController_report",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitReportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportReceiptDto"
                }
              }
            }
          }
        },
        "summary": "Report an approved review (Idempotency-Key required)",
        "tags": [
          "public-reviews"
        ]
      }
    },
    "/api/v1/admin/reviews": {
      "get": {
        "operationId": "ReviewsAdminController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "approved",
                "rejected",
                "spam"
              ]
            }
          },
          {
            "name": "businessId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repeatFlagged",
            "required": false,
            "in": "query",
            "description": "Only reviews flagged as repeat submissions",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "reported",
            "required": false,
            "in": "query",
            "description": "Only reviews with open abuse reports",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminReviewDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List reviews (status, business, repeatFlagged, reported)",
        "tags": [
          "admin-reviews"
        ]
      }
    },
    "/api/v1/admin/reviews/{id}": {
      "get": {
        "operationId": "ReviewsAdminController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminReviewDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin-reviews"
        ]
      }
    },
    "/api/v1/admin/reviews/{id}/approve": {
      "post": {
        "operationId": "ReviewsAdminController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModerateReviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminReviewDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin-reviews"
        ]
      }
    },
    "/api/v1/admin/reviews/{id}/reject": {
      "post": {
        "operationId": "ReviewsAdminController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModerateReviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminReviewDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin-reviews"
        ]
      }
    },
    "/api/v1/admin/reviews/{id}/spam": {
      "post": {
        "operationId": "ReviewsAdminController_spam",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModerateReviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminReviewDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin-reviews"
        ]
      }
    },
    "/api/v1/admin/reviews/{id}/redaction": {
      "patch": {
        "operationId": "ReviewsAdminController_redact",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedactReviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminReviewDto"
                }
              }
            }
          }
        },
        "summary": "Replace or restore the published text; the original and rating are never changed",
        "tags": [
          "admin-reviews"
        ]
      }
    },
    "/api/v1/admin/reports": {
      "get": {
        "operationId": "ReportsAdminController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "open",
                "investigating",
                "resolved"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminReportDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin-reports"
        ]
      }
    },
    "/api/v1/admin/reports/{id}": {
      "get": {
        "operationId": "ReportsAdminController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminReportDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin-reports"
        ]
      }
    },
    "/api/v1/admin/reports/{id}/investigate": {
      "post": {
        "operationId": "ReportsAdminController_investigate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestigateReportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminReportDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin-reports"
        ]
      }
    },
    "/api/v1/admin/reports/{id}/resolve": {
      "post": {
        "operationId": "ReportsAdminController_resolve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveReportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminReportDto"
                }
              }
            }
          }
        },
        "summary": "Record the outcome; removing content remains a separate review decision",
        "tags": [
          "admin-reports"
        ]
      }
    },
...
  },
  "info": {
    "title": "Melbourne Sphere API",
    "description": "REST API for the Melbourne Sphere directory and blog. Admin routes require the ms_admin_session cookie.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "adminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ms_admin_session"
      }
    },
    "schemas": {
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "PublicReviewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "rating": {
            "type": "number",
            "minimum": 1,
            "maximum": 5
          },
          "text": {
            "type": "string",
            "description": "Redacted text when an editor replaced it, otherwise the original"
          },
          "redacted": {
            "type": "boolean",
            "description": "True when an editor redacted the published text (SRS REV 003)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "displayName",
          "rating",
          "text",
          "redacted",
          "createdAt"
        ]
      },
      "SubmitReviewDto": {
        "type": "object",
        "properties": {
          "rating": {
            "type": "number",
            "minimum": 1,
            "maximum": 5,
            "description": "Whole number 1–5"
          },
          "displayName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80
          },
          "email": {
            "type": "string",
            "maxLength": 254,
            "writeOnly": true,
            "description": "Never published; used for moderation contact only"
          },
          "text": {
            "type": "string",
            "minLength": 20,
            "maxLength": 3000
          },
          "acknowledged": {
            "type": "boolean",
            "description": "Acknowledgement of the review guidelines and privacy notice (SRS PRIV 001)"
          },
          "captchaToken": {
            "type": "string",
            "description": "Turnstile token (SRS SEC 002)"
          },
          "website": {
            "type": "string",
            "description": "Honeypot: must stay empty",
            "writeOnly": true
          }
        },
        "required": [
          "rating",
          "displayName",
          "email",
          "text",
          "acknowledged"
        ]
      },
      "SubmissionReceiptDto": {
        "type": "object",
        "properties": {
          "receiptId": {
            "type": "string",
            "description": "Opaque reference the visitor can quote; not a review id"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending"
            ]
          },
          "message": {
            "type": "string",
            "description": "Neutral acknowledgement text"
          }
        },
        "required": [
          "receiptId",
          "status",
          "message"
        ]
      },
      "SubmitReportDto": {
        "type": "object",
        "properties": {
          "reviewId": {
            "type": "string",
            "description": "Approved review being reported; provide exactly one target (SRS REP 001)"
          },
          "commentId": {
            "type": "string",
            "description": "Approved comment being reported; provide exactly one target"
          },
          "reason": {
            "type": "string",
            "enum": [
              "spam",
              "offensive",
              "misleading",
              "privacy",
              "other"
            ]
          },
          "details": {
            "type": "string",
            "maxLength": 1000
          },
          "email": {
            "type": "string",
            "maxLength": 254,
            "writeOnly": true,
            "description": "Optional private contact; never disclosed"
          },
          "captchaToken": {
            "type": "string"
          },
          "website": {
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "reason"
        ]
      },
      "ReportReceiptDto": {
        "type": "object",
        "properties": {
          "receiptId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "received"
            ]
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "receiptId",
          "status",
          "message"
        ]
      },
      "AdminReviewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "businessId": {
            "type": "string"
          },
          "businessName": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "description": "Only present for moderators; never public"
          },
          "rating": {
            "type": "number"
          },
          "originalText": {
            "type": "string",
            "description": "Exactly as submitted (SRS REV 003)"
          },
          "publicText": {
            "type": "string",
            "nullable": true
          },
          "redactionReason": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected",
              "spam"
            ]
          },
          "moderationReason": {
            "type": "string",
            "nullable": true
          },
          "moderatorAdminId": {
            "type": "string",
            "nullable": true
          },
          "decidedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "repeatFlagged": {
            "type": "boolean",
            "description": "Same email reviewed this business recently (SRS REV 005)"
          },
          "openReportCount": {
            "type": "number"
          },
          "acknowledgedVersion": {
            "type": "string"
          },
          "version": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "businessId",
          "businessName",
          "displayName",
          "email",
          "rating",
          "originalText",
          "publicText",
          "redactionReason",
          "status",
          "moderationReason",
          "moderatorAdminId",
          "decidedAt",
          "repeatFlagged",
          "openReportCount",
          "acknowledgedVersion",
          "version",
          "createdAt"
        ]
      },
      "ModerateReviewDto": {
        "type": "object",
        "properties": {
          "expectedVersion": {
            "type": "number"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "description": "Recorded in the audit log; required for reject and spam"
          }
        },
        "required": [
          "expectedVersion"
        ]
      },
      "RedactReviewDto": {
        "type": "object",
        "properties": {
          "expectedVersion": {
            "type": "number"
          },
          "publicText": {
            "type": "string",
            "nullable": true,
            "maxLength": 3000,
            "description": "Published text; null restores the original"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "description": "Why the text was changed (SRS REV 003)"
          }
        },
        "required": [
          "expectedVersion",
          "publicText",
          "reason"
        ]
      },
      "AdminReportDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "targetType": {
            "type": "string",
            "enum": [
              "review",
              "comment"
            ]
          },
          "reviewId": {
            "type": "string",
            "nullable": true
          },
          "commentId": {
            "type": "string",
            "nullable": true
          },
          "parentId": {
            "type": "string",
            "description": "Business id for a review, article id for a comment"
          },
          "reason": {
            "type": "string"
          },
          "details": {
            "type": "string",
            "nullable": true
          },
          "reporterEmail": {
            "type": "string",
            "nullable": true,
            "description": "Only present for report handlers"
          },
          "targetSnapshot": {
            "type": "string",
            "description": "Public content as reported (SRS REP 001)"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "investigating",
              "resolved"
            ]
          },
          "outcome": {
            "type": "string",
            "nullable": true,
            "enum": [
              "retain",
              "remove",
              "spam"
            ]
          },
          "resolutionNote": {
            "type": "string",
            "nullable": true
          },
          "moderatorAdminId": {
            "type": "string",
            "nullable": true
          },
          "targetStatus": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected",
              "spam"
            ],
            "description": "Current status of the reported review or comment"
          },
          "version": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "resolvedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "targetType",
          "reviewId",
          "commentId",
          "parentId",
          "reason",
          "details",
          "reporterEmail",
          "targetSnapshot",
          "status",
          "outcome",
          "resolutionNote",
          "moderatorAdminId",
          "targetStatus",
          "version",
          "createdAt",
          "resolvedAt"
        ]
      },
      "InvestigateReportDto": {
        "type": "object",
        "properties": {
          "expectedVersion": {
            "type": "number"
          }
        },
        "required": [
          "expectedVersion"
        ]
      },
      "ResolveReportDto": {
        "type": "object",
        "properties": {
          "expectedVersion": {
            "type": "number"
          },
          "outcome": {
            "type": "string",
            "enum": [
              "retain",
              "remove",
              "spam"
            ],
            "description": "Removing the content is a separate review decision (SRS REP 002)"
          },
          "note": {
            "type": "string",
            "maxLength": 1000
          }
        },
        "required": [
          "expectedVersion",
          "outcome"
        ]
      },
...
    }
  }
}